fix(android-shell): sync AndroidShellFoundationTest + CrossClientVectorTest to current Rust FFI bridge shape (Sweep 50 B1)

+++ apps/android-shell/src/test/kotlin/org/mosaic/android/foundation/CrossClientVectorTest.kt
@@ -274,8 +274,8 @@ private class CorpusBackedGeneratedRustHeaderApi : GeneratedRustHeaderApi {
     }
     return RustHeaderParseFfiResult(
       code = RustHeaderStableCode.OK,
-      epochId = epochId,
-      shardIndex = shardIndex,
+      epochId = epochId.toLong(),
+      shardIndex = shardIndex.toLong(),
       tier = tier,
       nonce = nonce,
     )
